Add additional tests, parameterizedtests and the like

+++ spring-boot-unit-integration-testing/src/main/java/dev/aj/app/model/Grade.java
@@ -3,13 +3,13 @@ package dev.aj.app.model;
 public interface Grade {
     double getGrade();
 
-    int getId();
+    Long getId();
 
-    void setId(int id);
+    void setId(Long id);
 
-    int getStudentId();
+    Long getStudentId();
 
-    void setStudentId(int studentId);
+    void setStudentId(Long studentId);
 
     void setGrade(double grade);
 }
