+++ spring-boot-unit-integration-testing/src/main/java/dev/aj/app/model/MathGrade.java
@@ -24,9 +24,9 @@ public class MathGrade implements Grade{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-    private int studentId;
+    private Long studentId;
 
     private double grade;
 
@@ -36,22 +36,22 @@ public class MathGrade implements Grade{
     }
 
     @Override
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
 
     @Override
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     @Override
-    public int getStudentId() {
+    public Long getStudentId() {
         return this.studentId;
     }
 
     @Override
-    public void setStudentId(int studentId) {
+    public void setStudentId(Long studentId) {
         this.studentId = studentId;
     }
